requirements.txt
Flask
gunicorn
spacy
firebase-admin
google-auth
google-cloud-language==2.0.0
google-cloud-translate==2.0.1
google-cloud-storage
gcsfs
python-pptx==0.6.18
docx2txt==0.8
numpy
tldextract==3.1.0
beautifulsoup4==4.9.3